Separate live send startup factories
